Bug 1747305 - Couldn't regenerate the cert if the /etc/origin/logging is deleted
If the directory /etc/origin/logging is deleted and ansible is run with
redeploy-certificates.yml, all the files under the directory are
regenerated except passwd.yml.

This patch is adding the code to check if passwd.yml exists or not to
generate the file if it does not.

(cherry picked from commit bc1c86bc4ea6946aaefe12b0c0147d08ceac3e3d)